Organise button stories and offer all variables
