+++ cardigan/stories/components/Buttons/Buttons.stories.mdx
@@ -1,8 +1,7 @@
 import { Meta, Story } from '@storybook/addon-docs/blocks';
-import ControlReadme from '@weco/common/views/components/Control/README.md';
 import * as stories from './Buttons.stories.tsx';
 
-<Meta title={`Components/Buttons`} />
+<Meta title={`Components/Buttons: Basics`} />
 
 # Buttons
 
@@ -10,20 +9,6 @@ import * as stories from './Buttons.stories.tsx';
 
 <Story story={stories.buttonSolid} />
 
-## `Control`
+## `DropdownButton`
 
-<ControlReadme />
-
-  <Story story={stories.control} />
-
-## `ToolbarSegmentedControl`
-
-<Story story={stories.toolbarSegmentedControl} />
-
-## `DownloadButton`
-
-<Story story={stories.downloadButton} />
-
-## `FilterDropdown`
-
-<Story story={stories.filterDropdown} />
+<Story story={stories.dropdownButton} />
