+++ cardigan/stories/components/ButtonsAlternates/ButtonsAlternates.stories.mdx
@@ -1,0 +1,25 @@
+import { Meta, Story } from '@storybook/addon-docs/blocks';
+import ControlReadme from '@weco/common/views/components/Control/README.md';
+import * as stories from './ButtonsAlternates.stories.tsx'
+
+<Meta title={`Components/Buttons: Alternates`} />
+
+# Buttons with specific behaviours
+
+## `Control`
+
+<ControlReadme />
+
+  <Story story={stories.control} />
+
+## `ToolbarSegmentedControl`
+
+<Story story={stories.toolbarSegmentedControl} />
+
+## `DownloadButton`
+
+<Story story={stories.downloadButton} />
+
+## `FilterDropdown`
+
+<Story story={stories.filterDropdown} />
