+++ cardigan/stories/components/CaptionedImage/CaptionedImage.stories.mdx
@@ -1,4 +1,4 @@
-import { Meta, Canvas, Story, ArgsTable } from '@storybook/addon-docs/blocks';
+import { Meta, Story, ArgsTable } from '@storybook/addon-docs/blocks';
 import CaptionedImage from '@weco/content/components/CaptionedImage/CaptionedImage';
 import * as stories from './CaptionedImage.stories.tsx';
