+++ cardigan/stories/components/Contact/Contact.stories.mdx
@@ -9,7 +9,6 @@ import * as stories from './Contact.stories.tsx';
 
 <Readme />
 
-{' '}
 <Story story={stories.basic} />
 
 <ArgsTable />
