+++ common/views/components/Buttons/README.md
@@ -1,0 +1,14 @@
+## Purpose
+Button serves to offer one centralised component for all of our button types. Currently contains; 
+
+### Solid
+Default button and is a button element.
+
+### SolidLink
+Looks exactly like ButtonSolid but is a link and needs a `link` property
+
+### Dropdown
+Button element that acts like a Select element. Takes children for its modal content.
+
+
+[Edit this Readme on GitHub](https://github.com/wellcomecollection/wellcomecollection.org/edit/main/common/views/components/Buttons/README.md)
