+++ common/views/components/Control/README.md
@@ -2,4 +2,4 @@
 To allow a user to control the website UI in some way, e.g. 'Zoom' on the image viewer.
 
 
-[Edit this Readme on GitHub](https://github.com/wellcomecollection/wellcomecollection.org/edit/main/common/views/components/Buttons/Control/README.md)
+[Edit this Readme on GitHub](https://github.com/wellcomecollection/wellcomecollection.org/edit/main/common/views/components/Control/README.md)
